+++ src/Events/Integrations/Plugins/Tickets_Wallet_Plus/Apple_Wallet.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace TEC\Events\Integrations\Plugins\Tickets_Wallet_Plus;
+
+/**
+ * Class Apple_Wallet
+ * 
+ * @since TBD
+ * 
+ * @package TEC\Events\Integrations\Plugins\Tickets_Wallet_Plus
+ *
+ */
+class Apple_Wallet {
+	
+}
