+++ src/Events/Integrations/Provider.php
@@ -33,6 +33,6 @@ class Provider extends Service_Provider {
 		$this->container->register( Plugins\Colbri_Page_Builder\Provider::class );
 		$this->container->register( Plugins\Event_Tickets\Provider::class );
 		$this->container->register( Plugins\Elementor\Provider::class );
-		$this->container->register( Plugins\Event_Tickets_Wallet_Plus\Controller::class );
+		$this->container->register( Plugins\Tickets_Wallet_Plus\Controller::class );
 	}
 }
